+++ database/queries/categories.sql
@@ -1,9 +1,10 @@
 -- name: CreateCategory :one
 INSERT INTO categories (
-    name, 
+    user_id,
+    name,
     transaction_type_id
 ) VALUES (
-  $1, $2
+  $1, $2, $3
 )
 RETURNING *;
 
@@ -11,14 +12,16 @@ RETURNING *;
 SELECT *
   FROM categories
  WHERE id = $1
+   AND user_id = $2
  LIMIT 1;
 
 -- name: ListCategories :many
-SELECT * 
+SELECT *
   FROM categories
+ WHERE user_id = $1
  ORDER BY id
- LIMIT $1
-OFFSET $2;
+ LIMIT $2
+OFFSET $3;
 
 -- name: UpdateCategory :one
 UPDATE categories
@@ -26,8 +29,10 @@ UPDATE categories
        transaction_type_id = COALESCE(sqlc.narg('transaction_type_id'), transaction_type_id),
        updated_at = NOW()
  WHERE id = $1
+   AND user_id = sqlc.arg('user_id')
 RETURNING *;
 
 -- name: DeleteCategory :exec
 DELETE FROM categories
- WHERE id = $1;
+ WHERE id = $1
+   AND user_id = $2;
